refactor(core): Extract factory and finalize request-helpers barrel
