+++ packages/core/src/execution-engine/node-execution-context/utils/__tests__/outbound-user-agent.integration.test.ts
@@ -6,9 +6,12 @@ import { mock } from 'vitest-mock-extended';
 
 import type { ExecutionLifecycleHooks } from '@/execution-engine/execution-lifecycle-hooks';
 
-import { buildRfcStyleUserAgent, getDefaultN8nOutboundUserAgent } from '../outbound-user-agent';
 import { httpRequest } from '../request-helpers/http-request';
 import { proxyRequestToAxios } from '../request-helpers/legacy-request-adapter';
+import {
+	buildRfcStyleUserAgent,
+	getDefaultN8nOutboundUserAgent,
+} from '../request-helpers/outbound-user-agent';
 
 /** Exercises the full httpRequest → axios path for outbound User-Agent resolution. */
 describe('Outbound User-Agent (httpRequest integration)', () => {
